+++ src/main/kotlin/com/github/madwareru/intellijmdbook/project/MdBookProject.kt
@@ -1,0 +1,7 @@
+package com.github.madwareru.intellijmdbook.project
+
+data class MdBookProject(
+    val sourcesDirectory: String,
+    val buildDirectory: String,
+    val workingDirectory: String
+)