+++ src/main/kotlin/com/github/madwareru/intellijmdbook/project/MdBookProjectConstants.kt
@@ -1,0 +1,6 @@
+package com.github.madwareru.intellijmdbook.project
+
+object MdBookProjectConstants {
+    const val NAME = "MdBook"
+    const val MANIFEST_FILE = "book.toml"
+}