+++ src/main/kotlin/com/github/madwareru/intellijmdbook/project/services/MdBookProjectService.kt
@@ -1,0 +1,10 @@
+package com.github.madwareru.intellijmdbook.project.services
+
+import com.github.madwareru.intellijmdbook.project.MdBookProject
+import com.intellij.openapi.project.Project
+
+interface MdBookProjectService {
+    val project: Project
+    val mdBookProject: MdBookProject
+}
+