+++ src/main/kotlin/com/github/madwareru/intellijmdbook/vfs/ext.kt
@@ -1,0 +1,10 @@
+package com.github.madwareru.intellijmdbook.vfs
+
+import com.intellij.openapi.vfs.VirtualFile
+
+fun VirtualFile.findParent(searchedName: String ): VirtualFile? =
+    when {
+        parent == null -> null
+        parent.name == searchedName -> parent
+        else -> parent.findParent(searchedName)
+    }